ivshmem: add a register-only BAR
Add two new registers:
IRQSTATUS, to show the interrupt status without resetting it
IVSHMVER, to make sure guests can differentiate between ivshmem versions
... and reserve the rest for the future.

These registers are read-only at the moment. In the future, however,
there might be a need for writeable registers as well.

Current ivshmem version is 1, so as to guests which do not see BAR3
can use 0 internally.

Signed-off-by: Levente Kurusa <levex@linux.com>